+++ com.unity.mobile.android-logcat/CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.3-preview] - 2018-11-16
+
+### Minor fixes & Improvements.
+- Fixed the issue that package filter doesn't work correctly on devices below Android 7.
+- Fixed some issues at UI side.
+
 ## [0.1.2-preview] - 2018-10-30
 
 ### Minor fixes & Improvements.
 - Added documentation.
-- Fixed the issue that Android Logcat Package doesn't work with old devices like Android 4.4, 5.0 and 6.0.
+- Fixed the issue that Android Logcat Package doesn't work with devices below Android 7.
 
 ## [0.1.1-preview] - 2018-10-22